Evaluation code now use ground truth homography
